+++ src/main/index.ts
@@ -63,8 +63,8 @@ cli
     'Port pointing to a localhost dev server that serves your UI assets.',
   )
   .option(
-    '--signaling-url <url>',
-    'Url of the signaling server to use. By default, hc spin spins up a local development signaling server for you but this argument allows you to specify a custom one.',
+    '--relay-url <url>',
+    'Url of the relay server to use. By default, hc spin spins up a local development relay server for you but this argument allows you to specify a custom one.',
   )
   .option(
     '--force-admin-ports <ports>',
@@ -179,22 +179,22 @@ async function startLocalServices(): Promise<[string, string]> {
   const localServicesHandle = childProcess.spawn('kitsune2-bootstrap-srv');
   return new Promise((resolve) => {
     let bootStrapUrl;
-    let signalUrl;
+    let relayUrl;
     let bootstrapRunning = false;
-    let signalRunnig = false;
+    let relayRunning = false;
     localServicesHandle.stdout.pipe(split()).on('data', async (line: string) => {
       console.log(`[hc-spin] | [kitsune2-bootstrap-srv]: ${line}`);
       if (line.includes('#kitsune2_bootstrap_srv#listening#')) {
         const hostAndPort = line.split('#kitsune2_bootstrap_srv#listening#')[1].split('#')[0];
         bootStrapUrl = `http://${hostAndPort}`;
-        signalUrl = `ws://${hostAndPort}`;
+        relayUrl = `http://${hostAndPort}`;
       }
       if (line.includes('#kitsune2_bootstrap_srv#running#')) {
         bootstrapRunning = true;
-        signalRunnig = true;
+        relayRunning = true;
       }
-      if (bootstrapRunning && signalRunnig && bootStrapUrl && signalUrl)
-        resolve([bootStrapUrl, signalUrl]);
+      if (bootstrapRunning && relayRunning && bootStrapUrl && relayUrl)
+        resolve([bootStrapUrl, relayUrl]);
     });
     localServicesHandle.stderr.pipe(split()).on('data', async (line: string) => {
       console.log(`[hc-spin] | [hc run-local-services] ERROR: ${line}`);
@@ -211,7 +211,7 @@ async function spawnSandboxes(
   nAgents: number,
   happPath: string,
   bootStrapUrl: string,
-  signalUrl: string,
+  relayUrl: string,
   appId: string,
   networkSeed?: string,
   targetArcFactor?: number,
@@ -250,7 +250,7 @@ async function spawnSandboxes(
   if (targetArcFactor !== undefined) {
     generateArgs.push('--target-arc-factor', targetArcFactor.toString());
   }
-  generateArgs.push('--bootstrap', bootStrapUrl, 'webrtc', signalUrl);
+  generateArgs.push('--bootstrap', bootStrapUrl, 'quic', relayUrl);
   // console.log('GENERATE ARGS: ', generateArgs);
 
   let readyConductors = 0;
@@ -313,13 +313,13 @@ app.whenReady().then(async () => {
     );
   }
 
-  const [bootstrapUrl, signalingUrl] = await startLocalServices();
+  const [bootstrapUrl, relayUrl] = await startLocalServices();
 
   const [sandboxHandle, sandboxPaths, portsInfo] = await spawnSandboxes(
     CLI_OPTS.numAgents,
     happTargetDir ? happTargetDir : CLI_OPTS.happOrWebhappPath.path,
     CLI_OPTS.bootstrapUrl ? CLI_OPTS.bootstrapUrl : bootstrapUrl,
-    CLI_OPTS.singalingUrl ? CLI_OPTS.singalingUrl : signalingUrl,
+    CLI_OPTS.relayUrl ? CLI_OPTS.relayUrl : relayUrl,
     CLI_OPTS.appId,
     CLI_OPTS.networkSeed,
     CLI_OPTS.targetArcFactor,