+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 ### Fixed
 ### Changed
+- Holochain is now run with the iroh transport. The argument `--signaling-url` has been replaced with `--relay-url`.
 ### Removed
 
 ## 2026-01-13: v0.700.0-dev.0